96 - Refactor main.go
Many of the functions within main.go had more specialized packages
to host themselves, and were moved there. In addition, other changes
were made that modified the structure of several packages.

- To remove circular dependencies, many methods were moved to the github
package, as it handles much of the communication to and from github. This includes:

- Some of the above functions and structs were taken from template.go, as
they did not match the intended function of this package.

- TemplateInfo struct, methods, and tests added to template.go

- Created a new module, cli, to aggregate methods related to handling
cli input
